new-bot/requirements.txt
ccxt
pandas
numpy
python-dotenv
openai
